requirements.txt
# Core dependencies
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
pydantic>=2.0.0

# Computer vision and detection
opencv-python>=4.8.0
scikit-image>=0.21.0
scikit-learn>=1.3.0

# LLM integration
openai>=1.0.0
anthropic>=0.7.0

# UI
streamlit>=1.28.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0

